Refactor index functions
