--- src/db/kv/index/get.rs
@@ -1,42 +1,0 @@
-use crate::db::connectors::redis::get_redis_conn;
-use log::debug;
-use redis::JsonAsyncCommands;
-use serde::de::DeserializeOwned;
-use std::error::Error;
-
-/// Retrieves a JSON value from Redis at a specified path.
-///
-/// # Arguments
-///
-/// * `prefix` - A string slice that represents the prefix for the Redis key.
-/// * `key` - A string slice that represents the key under which the value is stored.
-/// * `path` - An optional string slice representing the JSON path from which the value should be retrieved. Defaults to the root path "$".
-///
-/// # Returns
-///
-/// Returns an `Option` containing the retrieved value if it exists, or `None` if it does not.
-///
-/// # Errors
-///
-/// Returns an error if the operation fails.
-pub async fn get<T: DeserializeOwned + Send + Sync>(
-    prefix: &str,
-    key: &str,
-    path: Option<&str>,
-) -> Result<Option<T>, Box<dyn Error + Send + Sync>> {
-    let mut redis_conn = get_redis_conn().await?;
-    let index_key = format!("{}:{}", prefix, key);
-    let json_path = path.unwrap_or("$").to_string(); // Ensure path is a String
-
-    // Use RedisJSON commands to get the value from the specified path
-    if let Ok(indexed_value) = redis_conn
-        .json_get::<String, String, String>(index_key.clone(), json_path)
-        .await
-    {
-        debug!("Restored key: {} with value: {}", index_key, indexed_value);
-        let value: Vec<T> = serde_json::from_str(&indexed_value)?;
-        return Ok(value.into_iter().next()); // Extract the first element from the Vec
-    }
-
-    Ok(None)
-}
