--- src/db/kv/index/get_set_range.rs
@@ -1,73 +1,0 @@
-use crate::db::connectors::redis::get_redis_conn;
-use std::error::Error;
-
-/// Retrieves a range of elements from a Redis set.
-///
-/// This function retrieves elements from a specified Redis set using a cursor-based approach.
-/// The range is defined by `skip` and `limit` parameters, where `skip` indicates the number
-/// of elements to skip and `limit` specifies the number of elements to retrieve.
-///
-/// # Arguments
-///
-/// * `prefix` - A string slice representing the prefix for the Redis keys.
-/// * `key` - A string slice representing the key under which the set is stored.
-/// * `skip` - The number of elements to skip.
-/// * `limit` - The number of elements to retrieve from the set after the skip.
-///
-/// # Returns
-///
-/// Returns a vector of strings containing the retrieved elements.
-///
-/// # Errors
-///
-/// Returns an error if the operation fails.
-pub async fn get_set_range(
-    prefix: &str,
-    key: &str,
-    skip: Option<usize>,
-    limit: Option<usize>,
-) -> Result<Option<Vec<String>>, Box<dyn Error + Send + Sync>> {
-    let mut redis_conn = get_redis_conn().await?;
-
-    let index_key = format!("{}:{}", prefix, key);
-    let mut cursor = "0".to_string();
-    let mut collected: Vec<String> = Vec::new();
-    let skip = skip.unwrap_or(0);
-    let limit = limit.unwrap_or(5);
-
-    let mut skipped = 0;
-
-    while collected.len() < limit {
-        let result: (String, Vec<String>) = redis::cmd("SSCAN")
-            .arg(&index_key)
-            .arg(&cursor)
-            .arg("COUNT")
-            .arg(limit)
-            .query_async(&mut redis_conn)
-            .await?;
-
-        let (new_cursor, items) = result;
-
-        for item in items {
-            if skipped < skip {
-                skipped += 1;
-                continue;
-            }
-            collected.push(item);
-            if collected.len() >= limit {
-                break;
-            }
-        }
-
-        cursor = new_cursor;
-        if cursor == "0" {
-            break; // End of the set reached
-        }
-    }
-
-    if collected.is_empty() {
-        Ok(None)
-    } else {
-        Ok(Some(collected))
-    }
-}
