+++ src/db/kv/index/mod.rs
@@ -1,18 +1,5 @@
-mod check_set_member;
-mod get;
-mod get_bool;
-mod get_list_range;
-mod get_set_range;
-mod set;
-mod set_list;
-mod set_multiple;
-mod set_set;
-
-pub use check_set_member::check_set_member;
-pub use get::get;
-pub use get_list_range::get_list_range;
-pub use get_set_range::get_set_range;
-pub use set::set;
-pub use set_list::set_list;
-pub use set_multiple::set_multiple;
-pub use set_set::set_set;
+/// Module for redis Indexing operations split into modules by Redis types
+pub mod json;
+pub mod lists;
+pub mod sets;
+pub mod sorted_sets;
