+++ src/models/post/details.rs
@@ -39,7 +39,7 @@ impl PostDetails {
         author_id: &str,
         post_id: &str,
     ) -> Result<Option<PostDetails>, Box<dyn std::error::Error + Send + Sync>> {
-        match Self::try_from_index(&[author_id, post_id]).await? {
+        match Self::try_from_index_json(&[author_id, post_id]).await? {
             Some(details) => Ok(Some(details)),
             None => Self::get_from_graph(author_id, post_id).await,
         }
@@ -71,7 +71,7 @@ impl PostDetails {
             Some(row) => {
                 let node: Node = row.get("p")?;
                 let post = Self::from_node(&node, author_id).await;
-                post.set_index(&[author_id, post_id]).await?;
+                post.put_index_json(&[author_id, post_id]).await?;
                 Ok(Some(post))
             }
             None => Ok(None),
